BE/pyproject.toml
[tool.poetry]
name = "soliton-day-ai-show"
version = "0.1.0"
description = ""
authors = ["sanjeev-manian"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.58.1"
rich = "^13.9.4"
streamlit = "^1.39.0"
tzlocal = "^5.2"
websockets = "^13.1"
sounddevice = "^0.5.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydub = "^0.25.1"
scipy = "^1.15.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
path = "C:\\sanjeev\\Training\\palmist\\"